requirements.txt
fastapi
uvicorn
python-multipart
pinecone-client
transformers
torch
Pillow